fix: move Tooltip to general components folder
