+++ src/components/Tooltip/Tooltip.js
@@ -2,7 +2,7 @@ import { Tooltip as UiTooltip } from '@dhis2/ui'
 import cx from 'classnames'
 import PropTypes from 'prop-types'
 import React from 'react'
-import styles from './styles/Tooltip.module.css'
+import styles from './Tooltip.module.css'
 
 const Tooltip = ({ disabled, content, children }) => (
     <UiTooltip
