+++ src/components/index.js
@@ -46,4 +46,5 @@ export { ResourcePickerField } from './ResourcePicker/ResourcePickerField.js'
 export { BasicOptions } from './BasicOptions/BasicOptions.js'
 export { SchemeContainer } from './ElementSchemes/SchemeContainer.js'
 export { StyledField } from './StyledField/StyledField.js'
+export { Tooltip } from './Tooltip/Tooltip.js'
 export { ValidationSummary } from './ValidationSummary/ValidationSummary.js'
