+++ src/pages/EarthEngineImport/EarthEngineImportForm.js
@@ -4,7 +4,7 @@ import { ReactFinalForm, Divider, Button } from '@dhis2/ui'
 import arrayMutators from 'final-form-arrays'
 import PropTypes from 'prop-types'
 import React, { useState, useContext, useRef } from 'react'
-import { Page, DataIcon } from '../../components/index.js'
+import { Page, DataIcon, Tooltip } from '../../components/index.js'
 import { FormAlerts, ImportButtonStrip } from '../../components/Inputs/index.js'
 import { TaskContext, getNewestTask } from '../../contexts/index.js'
 import { AssociatedGeometry } from './components/AssociatedGeometry.js'
@@ -15,7 +15,6 @@ import { MappingTable } from './components/MapGenderAgeGroupsTable.js'
 import { OrganisationUnits } from './components/OrganisationUnits.js'
 import { Periods } from './components/Periods.js'
 import { Rounding, defaultRoundingOption } from './components/Rounding.js'
-import { Tooltip } from './components/Tooltip.js'
 import styles from './EarthEngineImportForm.module.css'
 import { onImport } from './form-helper.js'
 import { getPeriods, getAggregations } from './util/earthEngineHelper.js'
